backend/requirements/base.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0
djangorestframework-simplejwt>=5.3,<6.0
django-mptt>=0.16,<1.0
django-environ>=0.11,<1.0
channels>=4.0,<5.0
channels-redis>=4.2,<5.0
psycopg2-binary>=2.9,<3.0
redis>=5.0,<6.0
reportlab>=4.1,<5.0
openpyxl>=3.1,<4.0
arabic-reshaper>=3.0,<4.0
python-bidi>=0.4,<1.0
gunicorn>=21.2,<23.0
whitenoise>=6.6,<7.0
